pkg/features: update filestructure
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
workspace_id: 7111cef4-3222-46ad-9b08-7e8b34e58cb1
