+++ api/pkg/features/enterprise/cloud/graphql/module.go
@@ -1,0 +1,9 @@
+package graphql
+
+import (
+	"getsturdy.com/api/pkg/di"
+)
+
+func Module(c *di.Container) {
+	c.Register(NewFeaturesRootResolver)
+}
