+++ api/pkg/features/enterprise/cloud/module.go
@@ -1,0 +1,10 @@
+package cloud
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/features/enterprise/cloud/graphql"
+)
+
+func Module(c *di.Container) {
+	c.Import(graphql.Module)
+}
