+++ api/pkg/features/enterprise/selfhosted/graphql/resolver.go
@@ -1,4 +1,4 @@
-package cloud
+package graphql
 
 import "getsturdy.com/api/pkg/graphql/resolvers"
 
@@ -12,7 +12,7 @@ func (r *FeaturesRootResolver) Features() []resolvers.Feature {
 	return []resolvers.Feature{
 		resolvers.FeatureBuildkite,
 		resolvers.FeatureGitHub,
-		resolvers.FeatureMultiTenancy,
-		resolvers.FeatureEmailAuth,
+		resolvers.FeatureLicense,
+		resolvers.FeaturePasswordAuth,
 	}
 }
