+++ api/pkg/features/graphql/module.go
@@ -1,0 +1,7 @@
+package graphql
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(NewFeaturesRootResolver)
+}
