--- api/pkg/features/graphql/module_cloud.go
@@ -1,13 +1,0 @@
-//go:build cloud
-// +build cloud
-
-package graphql
-
-import (
-	"getsturdy.com/api/pkg/di"
-	"getsturdy.com/api/pkg/features/graphql/enterprise/cloud"
-)
-
-func Module(c *di.Container) {
-	c.Register(cloud.NewFeaturesRootResolver)
-}
