--- api/pkg/features/graphql/module_oss.go
@@ -1,13 +1,0 @@
-//go:build !enterprise && !cloud
-// +build !enterprise,!cloud
-
-package graphql
-
-import (
-	"getsturdy.com/api/pkg/di"
-	"getsturdy.com/api/pkg/features/graphql/oss"
-)
-
-func Module(c *di.Container) {
-	c.Register(oss.NewFeaturesRootResolver)
-}
