+++ api/pkg/features/module/enterprise.go
@@ -1,0 +1,13 @@
+//go:build enterprise
+// +build enterprise
+
+package module
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/features/enterprise/selfhosted"
+)
+
+func Module(c *di.Container) {
+	c.Import(selfhosted.Module)
+}
